% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rAtte.R
\name{print.CMMRandomAttention}
\alias{print.CMMRandomAttention}
\title{Internal function.}
\usage{
\method{print}{CMMRandomAttention}(x, ...)
}
\arguments{
\item{object}{Class \code{CMMRandomAttention} objects.}
}
\description{
Internal function.
}
\keyword{internal}
