% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revealAtte.R
\name{print.ramchoiceRevealAtte}
\alias{print.ramchoiceRevealAtte}
\title{Internal function.}
\usage{
\method{print}{ramchoiceRevealAtte}(x, ...)
}
\arguments{
\item{object}{Class \code{ramchoiceRevealAtte} objects.}
}
\description{
Internal function.
}
\keyword{internal}
