% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revealPref.R
\name{print.ramchoiceRevealPref}
\alias{print.ramchoiceRevealPref}
\title{Internal function.}
\usage{
\method{print}{ramchoiceRevealPref}(x, ...)
}
\arguments{
\item{object}{Class \code{ramchoiceRevealPref} objects.}
}
\description{
Internal function.
}
\keyword{internal}
