% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revealPrefModel.R
\name{print.ramchoiceRevealPrefModel}
\alias{print.ramchoiceRevealPrefModel}
\title{Internal function.}
\usage{
\method{print}{ramchoiceRevealPrefModel}(x, ...)
}
\arguments{
\item{object}{Class \code{ramchoiceRevealPrefModel} objects.}
}
\description{
Internal function.
}
\keyword{internal}
