% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rAtte.R
\name{summary.CMMRandomAttention}
\alias{summary.CMMRandomAttention}
\title{Internal function.}
\usage{
\method{summary}{CMMRandomAttention}(object, ...)
}
\arguments{
\item{object}{Class \code{CMMRandomAttention} objects.}
}
\description{
Internal function.
}
\keyword{internal}
