% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revealAtte.R
\name{summary.ramchoiceRevealAtte}
\alias{summary.ramchoiceRevealAtte}
\title{Internal function.}
\usage{
\method{summary}{ramchoiceRevealAtte}(object, ...)
}
\arguments{
\item{object}{Class \code{ramchoiceRevealAtte} objects.}
}
\description{
Internal function.
}
\keyword{internal}
