% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revealPref.R
\name{summary.ramchoiceRevealPref}
\alias{summary.ramchoiceRevealPref}
\title{Internal function.}
\usage{
\method{summary}{ramchoiceRevealPref}(object, ...)
}
\arguments{
\item{object}{Class \code{ramchoiceRevealPref} objects.}
}
\description{
Internal function.
}
\keyword{internal}
