% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revealPrefModel.R
\name{summary.ramchoiceRevealPrefModel}
\alias{summary.ramchoiceRevealPrefModel}
\title{Internal function.}
\usage{
\method{summary}{ramchoiceRevealPrefModel}(object, ...)
}
\arguments{
\item{object}{Class \code{ramchoiceRevealPrefModel} objects.}
}
\description{
Internal function.
}
\keyword{internal}
